+++ src/components/chat-flow/chat-item-renderer/index.tsx
@@ -4,48 +4,49 @@ import { settingsStorage, chatEvent } from '@/services';
 import type { UiChatItem } from '../types';
 import TwoLinesMessage from '../two-lines-message';
 import SuperChatSticker from '../super-chat-sticker';
+import PinnedMessage from '../pinned-message';
 
 interface Props {
     chatItem: UiChatItem;
-    settings: settingsStorage.Settings;
+    messageSettings: settingsStorage.MessageSettings;
+    onClickClose?: React.MouseEventHandler;
 }
 
-const ChatItemRenderer: React.FC<Props> = ({ chatItem, settings }) => (
+const ChatItemRenderer: React.FC<Props> = ({
+    chatItem,
+    messageSettings,
+    onClickClose,
+}) => (
     <>
         {chatEvent.isSuperStickerItem(chatItem) && (
             <SuperChatSticker
                 chatItem={chatItem}
-                messageSettings={chatEvent.getMessageSettings(
-                    chatItem,
-                    settings,
-                )}
+                messageSettings={messageSettings}
             />
         )}
         {chatEvent.isNormalChatItem(chatItem) && (
             <TwoLinesMessage
                 chatItem={chatItem}
-                messageSettings={chatEvent.getMessageSettings(
-                    chatItem,
-                    settings,
-                )}
+                messageSettings={messageSettings}
             />
         )}
         {chatEvent.isSuperChatItem(chatItem) && (
             <TwoLinesMessage
                 chatItem={chatItem}
-                messageSettings={chatEvent.getMessageSettings(
-                    chatItem,
-                    settings,
-                )}
+                messageSettings={messageSettings}
             />
         )}
         {chatEvent.isMembershipItem(chatItem) && (
             <TwoLinesMessage
                 chatItem={chatItem}
-                messageSettings={chatEvent.getMessageSettings(
-                    chatItem,
-                    settings,
-                )}
+                messageSettings={messageSettings}
+            />
+        )}
+        {chatEvent.isPinnedItem(chatItem) && (
+            <PinnedMessage
+                chatItem={chatItem}
+                messageSettings={messageSettings}
+                onClickClose={onClickClose}
             />
         )}
     </>